Move to mjs. Use cfworker json-schema validator.
